demo/requirements.txt
# ActiveMirrorOS Demo Dependencies
# Minimal requirements for running the demo

pyyaml>=6.0
